\name{DeltaSimulatedPedigree}
\alias{DeltaSimulatedPedigree}
\docType{data}
\title{
Estimates of Jacquard Coefficients for Simulated SNP data according to a Pedigree
}
\description{
  A list object containing nine matrices with pairwise Jacquard coefficients
}
\usage{data("DeltaSimulatedPedigree")}
\format{A list with nine fields.}
\details{
  \code{DeltaSimulatedPedigree} can be generated by applying function \code{Jacquard.cls} to the
  data in \code{SimulatedPedigree}.
}
\references{
Graffelman, J., Weir, B.S. and Goudet, J. (2024) Estimation of Jacquard's genetic identity coefficients with
bi-allelic variants by constrained least-squares. Preprint at bioRxiv \doi{10.1101/2024.03.25.586682}.
}
\examples{
data(DeltaSimulatedPedigree)
}
\keyword{datasets}
