\name{GTC}
\alias{GTC}
\docType{data}
\title{
Joint Genotype Counts according to a Pedigree
}
\description{
  Contains a list object with nine fields, consisting of nine lower triangular matrices with the
  joint genotype counts for 111 individuals.
}
\usage{data("GTC")}

\format{A list with nine fields.}

\references{
Goudet, J. (2022) JGTeach: JG Teaching material. R package version 0.1.9. https://github.com/jgx65

Graffelman, J., Weir, B.S. and Goudet, J. (2024) Estimation of Jacquard's genetic identity coefficients with
bi-allelic variants by constrained least-squares. Preprint at bioRxiv \doi{10.1101/2024.03.25.586682}.
}

\examples{
data(GTC)
}

\keyword{datasets}
